refactor(validator): split Golden DKG ceremony tests
